domain-finder/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.2
djangorestframework==3.12.4
drf-api-logger==1.0.7
environ==1.0
idna==2.10
lxml==4.6.3
packaging==20.9
psycopg2==2.8.6
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
